requirements.txt
torch
torchvision
numpy
Pillow
segmentation-models-pytorch
matplotlib
